web_api/requirements.txt
redis==3.5.3
fastapi==0.73.0
uvicorn==0.17.5
rejson==0.5.6
timeloop==1.0.2
python-decouple==3.6

